remoce workflow from enclave
